LayoutTests/platform/ios/fast/canvas/offscreen-giant-transfer-to-imagebitmap-expected.txt

CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 16777216).
CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 16777216).
